New: Fix leaking asyncio loop listener on restart
After restarting a server, the associated asyncio loop was still
listening on the communication pipe between the processes. Since the
loop is re-use between restarts, the loop was permanently receiving an
unhandled "connection closed" message, which would cause the process to
use 100% CPU. We're now removing the listener upon restart.